More rearrangement to reduce diff in #4699
