+++ src/core/jsbind.h
@@ -1,0 +1,5 @@
+#include "jslib.h"
+#include "Python.h"
+
+PyObject*
+Js2Py_func_default_call_result(PyObject* self, JsVal jsresult, JsVal proxies);
