+++ src/core/jsproxy_call.c
@@ -1,65 +1,15 @@
 #include "Python.h"
 #include "error_handling.h"
 #include "js2python.h"
+#include "jsbind.h"
 #include "jslib.h"
 #include "pyproxy.h"
 #include "python2js.h"
-
-_Py_IDENTIFIER(create_future);
-_Py_IDENTIFIER(get_event_loop);
-_Py_IDENTIFIER(set_exception);
-_Py_IDENTIFIER(set_result);
-Js_IDENTIFIER(then);
-extern PyObject* asyncio_mod;
 
 Js_static_string(PYPROXY_DESTROYED_AT_END_OF_FUNCTION_CALL,
                  "This borrowed proxy was automatically destroyed at the "
                  "end of a function call. Try using "
                  "create_proxy or create_once_callable.");
-
-/**
- * Create a Future attached to the given Promise. When the promise is
- * resolved/rejected, the status of the future is set accordingly and
- * done_callback is called.
- */
-PyObject*
-wrap_promise(JsVal promise, JsVal done_callback)
-{
-  bool success = false;
-  PyObject* loop = NULL;
-  PyObject* set_result = NULL;
-  PyObject* set_exception = NULL;
-
-  PyObject* result = NULL;
-
-  loop = _PyObject_CallMethodIdNoArgs(asyncio_mod, &PyId_get_event_loop);
-  FAIL_IF_NULL(loop);
-
-  result = _PyObject_CallMethodIdNoArgs(loop, &PyId_create_future);
-  FAIL_IF_NULL(result);
-
-  set_result = _PyObject_GetAttrId(result, &PyId_set_result);
-  FAIL_IF_NULL(set_result);
-  set_exception = _PyObject_GetAttrId(result, &PyId_set_exception);
-  FAIL_IF_NULL(set_exception);
-
-  promise = JsvPromise_Resolve(promise);
-  FAIL_IF_JS_NULL(promise);
-  JsVal promise_handles =
-    create_promise_handles(set_result, set_exception, done_callback);
-  FAIL_IF_JS_NULL(promise_handles);
-  FAIL_IF_JS_NULL(JsvObject_CallMethodId(promise, &JsId_then, promise_handles));
-
-  success = true;
-finally:
-  Py_CLEAR(loop);
-  Py_CLEAR(set_result);
-  Py_CLEAR(set_exception);
-  if (!success) {
-    Py_CLEAR(result);
-  }
-  return result;
-}
 
 /**
  * Prepare arguments from a `METH_FASTCALL | METH_KEYWORDS` Python function to a
@@ -116,114 +66,6 @@ finally:
 }
 
 /**
- * This is a helper function for calling asynchronous js functions. proxies_id
- * is an Array of proxies to destroy, it returns a JsRef to a function that
- * destroys them and the result of the Promise.
- */
-EM_JS_VAL(JsVal, get_async_js_call_done_callback, (JsVal proxies), {
-  return function(result)
-  {
-    let msg = "This borrowed proxy was automatically destroyed " +
-              "at the end of an asynchronous function call. Try " +
-              "using create_proxy or create_once_callable.";
-    for (let px of proxies) {
-      Module.pyproxy_destroy(px, msg, false);
-    }
-    if (API.isPyProxy(result)) {
-      Module.pyproxy_destroy(result, msg, false);
-    }
-  };
-});
-
-// clang-format off
-EM_JS_VAL(JsVal, wrap_generator, (JsVal gen, JsVal proxies), {
-  proxies = new Set(proxies);
-  const msg =
-    "This borrowed proxy was automatically destroyed " +
-    "when a generator completed execution. Try " +
-    "using create_proxy or create_once_callable.";
-  function cleanup() {
-    proxies.forEach((px) => Module.pyproxy_destroy(px, msg));
-  }
-  function wrap(funcname) {
-    return function (val) {
-      if(API.isPyProxy(val)) {
-        val = val.copy();
-        proxies.add(val);
-      }
-      let res;
-      try {
-        res = gen[funcname](val);
-      } catch (e) {
-        cleanup();
-        throw e;
-      }
-      if (res.done) {
-        // Don't destroy the return value!
-        proxies.delete(res.value);
-        cleanup();
-      }
-      return res;
-    };
-  }
-  return {
-    get [Symbol.toStringTag]() {
-      return "Generator";
-    },
-    [Symbol.iterator]() {
-      return this;
-    },
-    next: wrap("next"),
-    throw: wrap("throw"),
-    return: wrap("return"),
-  };
-});
-
-EM_JS_VAL(JsVal, wrap_async_generator, (JsVal gen, JsVal proxies), {
-  proxies = new Set(proxies);
-  const msg =
-    "This borrowed proxy was automatically destroyed " +
-    "when an asynchronous generator completed execution. Try " +
-    "using create_proxy or create_once_callable.";
-  function cleanup() {
-    proxies.forEach((px) => Module.pyproxy_destroy(px, msg));
-  }
-  function wrap(funcname) {
-    return async function (val) {
-      if(API.isPyProxy(val)) {
-        val = val.copy();
-        proxies.add(val);
-      }
-      let res;
-      try {
-        res = await gen[funcname](val);
-      } catch (e) {
-        cleanup();
-        throw e;
-      }
-      if (res.done) {
-        // Don't destroy the return value!
-        proxies.delete(res.value);
-        cleanup();
-      }
-      return res;
-    };
-  }
-  return {
-    get [Symbol.toStringTag]() {
-      return "AsyncGenerator";
-    },
-    [Symbol.asyncIterator]() {
-      return this;
-    },
-    next: wrap("next"),
-    throw: wrap("throw"),
-    return: wrap("return"),
-  };
-});
-// clang-format on
-
-/**
  * __call__ overload for methods. Controlled by IS_CALLABLE.
  */
 PyObject*
@@ -235,7 +77,6 @@ JsMethod_Vectorcall_impl(JsVal target,
 {
   bool success = false;
   JsVal jsresult = JS_NULL;
-  bool destroy_args = true;
   PyObject* pyresult = NULL;
   JsVal proxies = JsvArray_New();
 
@@ -246,38 +87,13 @@ JsMethod_Vectorcall_impl(JsVal target,
   FAIL_IF_JS_NULL(jsargs);
   jsresult = JsvFunction_CallBound(target, thisarg, jsargs);
   FAIL_IF_JS_NULL(jsresult);
-  // various cases where we want to extend the lifetime of the arguments:
-  // 1. if the return value is a promise we extend arguments lifetime until the
-  //    promise resolves.
-  // 2. If the return value is a sync or async generator we extend the lifetime
-  //    of the arguments until the generator returns.
-  bool is_promise = JsvPromise_Check(jsresult);
-  bool is_generator = !is_promise && JsvGenerator_Check(jsresult);
-  bool is_async_generator =
-    !is_promise && !is_generator && JsvAsyncGenerator_Check(jsresult);
-  destroy_args = (!is_promise) && (!is_generator) && (!is_async_generator);
-  if (is_generator) {
-    jsresult = wrap_generator(jsresult, proxies);
-  } else if (is_async_generator) {
-    jsresult = wrap_async_generator(jsresult, proxies);
-  }
-  FAIL_IF_JS_NULL(jsresult);
-  if (is_promise) {
-    // Since we will destroy the result of the Promise when it resolves we deny
-    // the user access to the Promise (which would destroyed proxy exceptions).
-    // Instead we return a Future. When the promise is ready, we resolve the
-    // Future with the result from the Promise and destroy the arguments and
-    // result.
-    pyresult = wrap_promise(jsresult, get_async_js_call_done_callback(proxies));
-  } else {
-    pyresult = js2python(jsresult);
-  }
+  pyresult = Js2Py_func_default_call_result(NULL, jsresult, proxies);
   FAIL_IF_NULL(pyresult);
 
   success = true;
 finally:
   Py_LeaveRecursiveCall(/* " in JsMethod_Vectorcall" */);
-  if (!success || destroy_args) {
+  if (!success) {
     // If we succeeded and the result was a promise then we destroy the
     // arguments in async_done_callback instead of here. Otherwise, destroy the
     // arguments and return value now.
@@ -286,8 +102,6 @@ finally:
       JsvArray_Push(proxies, jsresult);
     }
     destroy_proxies(proxies, &PYPROXY_DESTROYED_AT_END_OF_FUNCTION_CALL);
-  } else {
-    gc_register_proxies(proxies);
   }
   if (!success) {
     Py_CLEAR(pyresult);
